requirements.txt
groq==0.9.0
yfinance==0.2.40
gradio==4.31.0
plotly==5.22.0
requests==2.31.0
python-dotenv==1.0.1
asyncio==3.4.3
pandas==2.2.2
pydantic==2.10.3
tenacity==8.3.0
